fix: routing renamed files properly
